I hope to god this fixes bugs or at least shows where they are

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/hardware/mechanisms/auxiliary/ITurret.java
@@ -3,8 +3,8 @@ package org.firstinspires.ftc.teamcode.hardware.mechanisms.auxiliary;
 import org.firstinspires.ftc.teamcode.core.hardware.state.Component;
 import org.firstinspires.ftc.teamcode.core.hardware.state.StatefulMechanism;
 
-public interface ITurret extends Component, StatefulMechanism<Integer> {
-  void turnToDegrees(int degrees);
+public interface ITurret extends Component, StatefulMechanism<Double> {
+  void turnToDegrees(double degrees);
 
   void turnToFront();
